refactor: Move IterableCounter to a dedicated namespace
In #2708 and likely upcoming PRs, we will have more `iterable` related
utility, so I moving this to a dedicated namespace for better
organisation.
