+++ src/Framework/Iterable/IterableCounter.php
@@ -33,9 +33,10 @@
 
 declare(strict_types=1);
 
-namespace Infection;
+namespace Infection\Framework\Iterable;
 
 use function count;
+use Infection\CannotBeInstantiated;
 use Infection\Console\OutputFormatter\AbstractOutputFormatter;
 use function is_array;
 use function iterator_to_array;
